refactor: demo 페이지에 코로케이션 (- prefix 활용)
