+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './CommonLoader';
 export * from './Button';
-export * from './Grid';
