+++ src/routeTree.gen.ts
@@ -11,21 +11,21 @@
 // Import Routes
 
 import { Route as rootRoute } from './routes/__root'
-import { Route as DragDemoImport } from './routes/drag-demo'
 import { Route as AboutImport } from './routes/about'
+import { Route as DragDemoRouteImport } from './routes/drag-demo/route'
 import { Route as IndexImport } from './routes/index'
 
 // Create/Update Routes
 
-const DragDemoRoute = DragDemoImport.update({
-  id: '/drag-demo',
-  path: '/drag-demo',
-  getParentRoute: () => rootRoute,
-} as any)
-
 const AboutRoute = AboutImport.update({
   id: '/about',
   path: '/about',
+  getParentRoute: () => rootRoute,
+} as any)
+
+const DragDemoRouteRoute = DragDemoRouteImport.update({
+  id: '/drag-demo',
+  path: '/drag-demo',
   getParentRoute: () => rootRoute,
 } as any)
 
@@ -46,18 +46,18 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof IndexImport
       parentRoute: typeof rootRoute
     }
+    '/drag-demo': {
+      id: '/drag-demo'
+      path: '/drag-demo'
+      fullPath: '/drag-demo'
+      preLoaderRoute: typeof DragDemoRouteImport
+      parentRoute: typeof rootRoute
+    }
     '/about': {
       id: '/about'
       path: '/about'
       fullPath: '/about'
       preLoaderRoute: typeof AboutImport
-      parentRoute: typeof rootRoute
-    }
-    '/drag-demo': {
-      id: '/drag-demo'
-      path: '/drag-demo'
-      fullPath: '/drag-demo'
-      preLoaderRoute: typeof DragDemoImport
       parentRoute: typeof rootRoute
     }
   }
@@ -67,42 +67,42 @@ declare module '@tanstack/react-router' {
 
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
+  '/drag-demo': typeof DragDemoRouteRoute
   '/about': typeof AboutRoute
-  '/drag-demo': typeof DragDemoRoute
 }
 
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
+  '/drag-demo': typeof DragDemoRouteRoute
   '/about': typeof AboutRoute
-  '/drag-demo': typeof DragDemoRoute
 }
 
 export interface FileRoutesById {
   __root__: typeof rootRoute
   '/': typeof IndexRoute
+  '/drag-demo': typeof DragDemoRouteRoute
   '/about': typeof AboutRoute
-  '/drag-demo': typeof DragDemoRoute
 }
 
 export interface FileRouteTypes {
   fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths: '/' | '/about' | '/drag-demo'
+  fullPaths: '/' | '/drag-demo' | '/about'
   fileRoutesByTo: FileRoutesByTo
-  to: '/' | '/about' | '/drag-demo'
-  id: '__root__' | '/' | '/about' | '/drag-demo'
+  to: '/' | '/drag-demo' | '/about'
+  id: '__root__' | '/' | '/drag-demo' | '/about'
   fileRoutesById: FileRoutesById
 }
 
 export interface RootRouteChildren {
   IndexRoute: typeof IndexRoute
+  DragDemoRouteRoute: typeof DragDemoRouteRoute
   AboutRoute: typeof AboutRoute
-  DragDemoRoute: typeof DragDemoRoute
 }
 
 const rootRouteChildren: RootRouteChildren = {
   IndexRoute: IndexRoute,
+  DragDemoRouteRoute: DragDemoRouteRoute,
   AboutRoute: AboutRoute,
-  DragDemoRoute: DragDemoRoute,
 }
 
 export const routeTree = rootRoute
@@ -116,18 +116,18 @@ export const routeTree = rootRoute
       "filePath": "__root.tsx",
       "children": [
         "/",
-        "/about",
-        "/drag-demo"
+        "/drag-demo",
+        "/about"
       ]
     },
     "/": {
       "filePath": "index.tsx"
     },
+    "/drag-demo": {
+      "filePath": "drag-demo/route.tsx"
+    },
     "/about": {
       "filePath": "about.tsx"
-    },
-    "/drag-demo": {
-      "filePath": "drag-demo.tsx"
     }
   }
 }
