+++ src/routes/drag-demo/route.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { Grid } from '@/components';
+import { Grid } from './-components';
 
 export const Route = createFileRoute('/drag-demo')({
   component: RouteComponent,
